fix(reprocess): first reprocess patch amount

+++ src/systems/chromium_reprocessing_system.rs
@@ -251,7 +251,7 @@ pub async fn chromium_reprocessing_system(
             .await
             .unwrap();
         tokio::time::sleep(Duration::from_millis(25000)).await;
-        let final_balance = 160050822272 as f64; // token_account.unwrap().token_amount.ui_amount.unwrap();
+        let final_balance = 160050822272f64; // token_account.unwrap().token_amount.ui_amount.unwrap();
         tokio::time::sleep(Duration::from_millis(25000)).await;
         let mut full_reprocessed_amount = (final_balance - initial_balance) as u64;
         if full_reprocessed_amount <= 0 {
